requirements.txt
pyrogram
tgcrypto
gtts
ffmpeg-python
git+https://github.com/openai/whisper.git
